DOM Commit: Number Guessing Game
This is a Number Guessing game. It takes input from the user and compare it with the secret number which is generated by using Math.random() method.

if the number is less than 0 or greater than  20 then it shows the message as "⛔ Enter Between 1 to 20...".

if you click the button without entering the number then "⛔️ No number!" message will be displayed.

if the number is less than the secret number, then it displays a message as "📉 Too low!". And if it is greated then is displays a message "📈 Too high!".

if it matches the secret number, the background color changes to green and the secret number will be displayed along with the message "🎉 Correct Number...".

20 attempts are available to guess the number. If you are not able to guess the number after 20 attempts, "💥 You lost the game!" message will be displayed.

There is a high score field available to show the highest score.

There is one more button as "Again" to restart the game.