Pig Game Commit
This is the HTML, CSS and Javascript project. It is a two player Game.

At initial stage Player 1 will roll the dice and he will get a number between 1 to 6. He can roll the dice multiple times, Once he get 1 then he will lost the current score and the another player will get a chance.
Player can save his score to the final score by clicking on hold button. by doing this, the current score will get added to the final score and it becomes 0.
To win this game, the player need to get 100 score before the other player. Once the player wins the game stops, and the background color of the winning player gets changed

There are 3 buttons on the screen:
1. 🔄 New game
2. 🎲 Roll dice
3. 📥 Hold

1. New Game Button: By clicking on this button all the score will reset and dice will hide and the game starts from the begining.

2. Roll Dice Button: By clickin on this button, the dice will roll and the number will get added to the active players current score.
There are basically 2 scores:

1. Current Score: It is the score which is calculated by adding the number the active player is getting by rolling the dice. the current score becomes 0 when the active player gets 1 or it becomes 0 when player clicks on hold button to save the current score to his final score.

2. Final Score: This is the score, which is calculated when the player clicks on Hold Button. When it becomes 100 the player wins the game.

3. Hold Button: By clicking on this button the active player can add his current score to his final score and the current score becomes 0 and the other become active to play the game.

-------------------------------------------

On Technical side:

I have used:
1. HTML, CSS and Javascript to build this game.
2. Used if and else statements to build various logics.
3. Used array to store players final score
4. Used classList on DOM Element to add and remove dice based on the conditions.
5. Targeted each HTML Elements in JS with the help of Class and Id selector.
6. Used Event Listener and Event Handler to perform logics on each Button.
7. Created generic function and used it multiple times in different buttons.